fix: Update the state struct
We want to read the genesis block and genesis state from
duna/jamtestnet. Therefore, this changes updates our state struct to
match to latest data structure.

However, when I tried to read the genesis state binary file, I found
that the accounts information could not be parsed correctly.
I've submitted an issue to jamtestnet repository, and they will update
the test data in next version.

Updates in this commit:

- Implemented an unmarshal JSON function for the state struct
- Implemented an encode function for the state struct
- Implemented a decode function for the state struct

For now, I am ignoring the delta(accounts) property when running the test.